requirements.txt
flask
Werkzeug
jinja2
flask-login
flask_migrate
WTForms
flask_wtf
flask-sqlalchemy
sqlalchemy
email_validator
flask-restx
python-dotenv
gunicorn
Flask-Minify
flask-dance
blinker
psycopg2-binary
sqlfluff
openai
sql-metadata
sql-formatter
markdown
pymdown-extensions>=10.0
pygments>=2.19.2
sqlparse
sqlglot
hvac
pyyaml